+++ Lab3/PhongVrtShader.glsl
@@ -1,6 +1,6 @@
-#version 430
+//#version 430
 
-precision highp float;
+precision mediump float;
 
 attribute vec3 position;
 attribute vec3 normal;
